FIX: sgb-tree default pointer size upgrade to 64 bit (#462)

+++ src/dbzero/core/collections/SGB_Tree/SGB_CompressedLookupTree.hpp
@@ -144,7 +144,7 @@ DB0_PACKED_END
         typename ItemEqualType,
         typename CompressedItemEqualType,
         typename CapacityType = std::uint16_t, 
-        typename AddressType = std::uint32_t, 
+        typename AddressType = std::uint64_t,
         typename HeaderType = db0::o_null,
         typename TreeHeaderType = db0::o_null>
     class sgb_compressed_lookup_types
